requirements.txt
streamlit
opencv-python-headless
pytesseract
pandas
numpy
Pillow
